Update pmeg.c
Added some missing headers; removed comment of old file name. Was originally in C++ in VS2019, but decided to switch to code blocks. That way the project isn't tied to any license changes in the future (who knows what MS will do in the future under the leadership of another CEO)